Bump version
Fixes from testing
